scaffolding for more notation stuff
big refactor is going to be needed to get everything pointing at the
notation module(s) instead of a random assortment of `0oXY` and
string-notation and all that. Ideally this will eliminate the need for
the NOTATION_TO_INDEX &c tables in `piece_lookup.rs` and make the
notation modules the primary way to talk about notation.

Hopefully they'll all be able to const-time compile to indicies and the
methods can all rely on the `Square` struct which is the simplest thing
and also the easiest to constify.