the beginning of the great reorganization
I've shuffled things around and disabled some stuff. I'm going to
deprecate Ply (and thus a lot of Game as well), and continue to build up
the design I've been working towards with the improved notation
management.

Next there is refactoring the code to use the notation system
consistently, and building up the various notation types.