mutation run + tuning for different VM
My big machine needs some work before it's usable for dev again, and
TBH, I'm thinking I might not develop directly on it, as it's a bit
inconvenient rn. I'm moving it's root disk to iSCSI (hopefully) and that
should improve ergonomics a bit (no more netboot builds for now) and buy
me time to figure out the netboot thing once and for all.