incremental refactoring towards new Play trait
This steps towards a new version of the Play trait that embeds some of
the stuff from `ChessAction` directly into the trait. Ultimately I
*think* should be able to eliminate a lot of the assumptions around what
the "Move"/"BEN" assumption and rely on traits, but for the moment this
is easier to get the functionality in place, then I can genericize it
later.