little stub for how drivers should eventually work
this is mostly just me working out an API, but I think this is going to
be what I use more generically. I need to still experiment with how
state update will work, but this gives me a very extensible messaging
system, and if I collapse the output to an enum it should be easy for
clients to filter which variant(s) they care about. I suppose a similar
trick may be possible w/ a boxed response, but I think that'd make the
client end much more complicated.