basic zobrist caching in place
I have not really done any testing, but it still counts the same values
for perft 3 and 4, and it's much quicker now (8s -> 2s for perft 3, 180s
-> 20s for perft 4.

It doesn't do incremental update really yet, in principle the position
should maintain it's zobrist incrementally, and ideally the cache would
handle calculating the board state and position would just manage it's
alterations (using the cache to do the metadata update, etc).

This would also allow `unmake` to be a bit cleverer.

I haven't done any kind of collision checking either, right now it will
happily clobber without any real way of knowing.

I also need to centralize the Cache in the MoveGen so multiple
`Position`s can share it, for that I'll probably need to either give
`Position` a handle late in the process, or otherwise wrap positions in
some handle so they can be cache-aware.